REFACTOR: Don't include prefix in file names
We no longer include the "projects/projectname/" prefix when dealing
with project files, as the information about which project a file
belongs to is already available. The uploadFile function still returns
the file with the prefix however.

Although The return value does not appear to be read, remove the prefix
to make the behaviour consistent with the rest of the platform.