FEAT: Add method-level throttling to APIs
Both sVEP and sBeacon APIs have no DDoS protections against otherwise
valid requests. This means that a flood of valid requests can cause
downstream cost blowouts and potentially undefined behaviour.

Limit the rate at which each API endpoint can be called. This means that
if a given endpoint is called at a rate greater than 100/s for long
enough (such that the queue of waiting requests exceeds 1000) future
requests to the same endpoint will be rejected with a 429 Too Many
Requests Error. This will impact all clients, and so a DDoS that
triggers this will actually deny the service for the duration of the
attack. However, the resource costs of such an attack will be limited to
100 requests per second. These numbers can be changed by altering the
relevant variables using terraform.tfvars.