FEAT: Reduce presigned upload expiry to 5 minutes
The presigned upload url expires after 1 hour, but as long as the upload
begins before expiry, the upload can run to completion.

Reduce the time before the upload url expires to 5 minutes. Of note,
this works for the post url, but may not work if we try to use multipart
uploads, as a final api call must be made after the last part completes,
which would likely be rejected if the expiration time has passed. Also
it appears if an upload is attempted after the expiry time, the
permission denied error is only returned after the upload has completed
(the file is not saved to s3).