FIX: Remove resource churn on using count for svep
When introducing count for the sVEP module, terraform attempts to delete
all the sVEP resource and recreate them, causing lost reference data and
recreating cloudfront distributions.

Use the terraform moved block to inform terraform of the changed module
location, so it is able to instead update the state locations of each
resource, rather than the resources themselves.