FEAT: Pull software version from latest commit
The software version displayed on the frontend footer is just a
placeholder.

Use git describe to populate the software version in the footer. This
value will be the latest tag, marking the presence of extra commits as
(-<number_of_commits>-<latest_commit_sha>) and the presence of
uncommitted changes (a "dirty" environment) with a "+" suffix.