FEAT: Add error_catcher module
When errors occur in pipelines, often an error is returned to the user.
But in the case of an uncaught error, the lambda function fails and
nobody is notified. This does not meet requirements.

Add a cloudwatch alarm to trigger when there are errors in lambda
functions. This alarm triggers a lambda function to find the errors and
send an email to the gaspi-admin-email. The function also starts a
schedule to call itself every 20 minutes while the alarm is present, as
the alarm only triggers a function when changing states. The function
looks at logs around the time of the error metric, as well as 16 minutes
into the future, to account for the fact that that the metric measures
from the invocation time, not the error time, and rounds the timestamp
down to the start of the period (in this case 60 seconds). Match log and
metric errors to seek out any errors that are not covered by log
messages, and track recent metric and log errors in S3 to avoid sending
notification emails for an error more than once.

The notification emails contain all new recent errors and counts,
grouped by error message (with some fuzziness to account for error
messages that differ only in unimportant details, like requestId), and
link to view the first of each such error in cloudwatch. If no log error
is found for an error metric, return a suitable message and a link to
the log group.