FIX: Allow additional error kinds to be caught
When a function is sampled for Xray, the final summary line in the logs
is multiple lines. This causes the timeout pattern to miss, as the final
$ character does not correctly match the newline.

Remove the requirement for "Status: timeout" to be at the end of the
line, to pick up log messages that contain additional information.