FIX: Send error emails with configuration sets
If a custom SES configuration-set is present, ses.send_email will
attempt to use it, but lacks permissions to do so.

Add permissions to use configuration sets, so if one is selected by
default, or the sending identity has one set, the error emails are still
able to go through.