FIX: Mark pgxflow module as moved to counted
pgxflow was once a single module, but has since been changed to a
count-defined module, to facilitate easy switching on and off. However
terraform needs to be informed that the underlying resources are the
same, and only their location has moved.

Mark the module as having moved from the old location to the new. This
is only required in order to upgrade existing infrastructure, and has no
effect on new deployments.